fix: game exception implementation placements
